check for sessions
