+++ client/pages/LoginPage1.html
@@ -6,7 +6,7 @@
   <title>Auth0 SDK Test</title>
 </head>
 <body>
-  <h1>Freelancer Platform Auth Example</h1>
+  <h1>Freelancer Platform Auth Example</h1> 
   <p>Welcome to the login page</p>
 
   <button id="btn-login" disabled="true" onclick="login()">Log in</button>
