+++ client/scripts/auth_config.json
