+++ client/scripts/config.env
@@ -1,0 +1,4 @@
+PORT = 3000
+GOOGLE_CLIENT_ID = 438436633528-nbqicia324dp59qced3vtlr684g2v69s.apps.googleusercontent.com
+GOOGLE_CLIENT_SECRET = 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
+GOOGLE_CLIENT_WEBSITE = https:/localhost:3000/
