+++ package.json
@@ -22,11 +22,15 @@
   "homepage": "https://github.com/technicalGabs/Freelancerio#readme",
   "dependencies": {
     "@auth0/auth0-spa-js": "^2.1.3",
+    "@googleapis/docs": "^3.3.3",
+    "cors": "^2.8.5",
     "dotenv": "^16.5.0",
     "express": "^5.1.0",
     "express-oauth2-jwt-bearer": "^1.6.1",
+    "googleapis": "^148.0.0",
     "mongoose": "^8.13.2",
-    "cors": "^2.8.5"
+    "passport": "^0.7.0",
+    "passport-google-oauth": "^2.0.0"
   },
   "devDependencies": {
     "autoprefixer": "^10.4.21",
